# swagger.json
{
  "openapi": "3.1.0",
  "servers": [
    {
      "url": "{scheme}://{host}/api/v1",
      "variables": {
        "scheme": {
          "enum": [
            "https",
            "http"
          ],
          "default": "https"
        },
        "host": {
          "default": "localhost:3000"
        }
      }
    },
    {
      "url": "/api/v1"
    }
  ],
  "info": {
    "version": "0.0.1",
    "title": "mycelia-api",
    "description": "",
    "license": {
      "name": "Apache-2.0"
    }
  },
  "paths": {
...
    "/podcast/config": {
      "get": {
        "tags": [
          "podcast"
        ],
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "x-samchon-accessor": [
          "podcast",
          "config",
          "getPodcastConfig"
        ],
        "x-samchon-controller": "PodcastConfigController"
      },
      "put": {
        "tags": [
          "podcast"
        ],
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {}
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "x-samchon-accessor": [
          "podcast",
          "config",
          "updatePodcastConfig"
        ],
        "x-samchon-controller": "PodcastConfigController"
      }
    },
...
  },
  "components": {
...
    "securitySchemes": {
      "bearer": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
...
}